🎬 Genre not assigned exception

+++ app/movies/actions/remove_movie_genre.py
@@ -1,5 +1,6 @@
 from uuid import UUID
 
+from app.movies.domain.exceptions import GenreNotAssignedException
 from app.movies.domain.repositories.movie_repository import MovieRepository
 
 
@@ -8,4 +9,9 @@ class RemoveMovieGenre:
         self._repository = repository
 
     def execute(self, movie_id: UUID, genre_id: UUID) -> None:
+        movie = self._repository.get(id=movie_id)
+
+        if not movie.has_genre(genre_id=genre_id):  # type: ignore
+            raise GenreNotAssignedException()
+
         self._repository.remove_genre(movie_id, genre_id)
